let the profiles of zsh and bash export GTK_IM_MODULE etc. too
Ok, I admit that these things are only GUI related but one would always
start GUI program from a terminal. If we don't export them in the
profiles of zsh, bash and etc., they will be overwritten by system
settings.

But this commit violate the Do-Not-Repeat-Yourself principle. I may
perfect that peace later.