mv shell related stuff to shells/ and add shells/profile
Bash won't source ~/.profile of it found ~/.bash_profile. Zsh won't
source it either when ~/.zprofile was found. But they all source it
instead of their own startup files when invoked as sh. So I want to put
the common stuff their and let .bash_profile and .zprofile source it
manually.