zsh.vim
Fix the bug that if there is a ' sign in a comment, highlight will break
when that line is absent on the screen.

After some `hard` debuging, I found that it is the 'display' option
that should be blamed. According to the help, 'If the "display" argument
is given,  this item will be skipped when the detected highlighting will
not be displayed.'. So if the line is absent on current screen, it won't
be parsed at all... That means the ' sign in the line will trigger the
zshString syntax. You will got the whole file highlighted as strings
until an other ' sign appears...